feat: Implement Redux middleware.
After hacking at Stately as a layer on top of Redux it became apparent that
it would ultimately become Redux with a little state machine sugar on top. I
decided to give using Redux middleware a try and it is much simpler while providing the same
enforcement.

Typescript has also been removed in favor of raw JS -- implementing Redux middleware
in Typescript is frustrating.